+++ src/tpl/wrap-default.php
@@ -1,3 +1,10 @@
+<?php
+/**
+ * WordPress admin 'wrap' div
+ * 
+ * @var WPHelper\AdminPage $this
+ */
+?>
 <div class="wrap">
 	<h1><?= get_admin_page_title() ?></h1>